utils/adso/requirements.txt
beautifulsoup4==4.4.1
cssutils==1.0.2
deepspeech==0.1.1
enum34==1.1.6
future==0.16.0
lxml==4.1.1
numpy==1.14.1
progressbar2==3.35.2
python-utils==2.3.0
scipy==1.0.0
six==1.11.0
